requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
twilio
requests
python-multipart